Rename status.py to daemon_status.py
